+++ CHANGES.md
@@ -5,7 +5,9 @@ Changes in mantohtml
 v2.0.1 - YYYY-MM-DD
 -------------------
 
+- Added `--section` option to specify "section" metadata.
 - Fixed a buffer overrun bug (Issue #1)
+- Fixed the default page title to be `NAME(SECTION)`.
 
 
 v2.0.0 - 2022-10-09